fix: serialize empty concept to none
