+++ tests/arches_django/test_arches_django.py
@@ -1,5 +1,6 @@
 import pytest
 import json
+from fixtures import person_ash, person_ashs
 
 JSON_PERSON = """
 {
@@ -54,20 +55,6 @@ JSON_PERSON = """
 }
 """
 
-
-@pytest.fixture
-def person_ash(arches_orm):
-    Person = arches_orm.models.Person
-    person = Person.create()
-    ash = person.name.append()
-    ash.full_name = "Ash"
-    return person
-
-@pytest.fixture
-def person_ashs(arches_orm, person_ash):
-    person_ash.save()
-    yield person_ash
-    person_ash.delete()
 
 @pytest.mark.django_db
 def test_can_save_with_name(arches_orm):
